requirements.txt
streamlit==1.29.0
gradio==4.8.0
torch==2.1.0
torchvision==0.16.0
transformers==4.36.0
git+https://github.com/openai/CLIP.git
hnswlib==0.8.0
Pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
matplotlib==3.7.2
seaborn==0.12.2
pandas==1.5.3
tqdm==4.66.1
sentence-transformers==2.2.2
google-generativeai==0.3.2
python-dotenv==1.0.0
scikit-learn==1.3.0
plotly==5.17.0
streamlit-image-coordinates==0.1.6 